feat : websocket

+++ Test/mockup.py
@@ -15,7 +15,8 @@ PUBLISH_INTERVAL = 1 / PUBLISH_HZ
 
 BASE_TOPIC = "vehicle/car_01"
 
-can0 ={        "latest" : {
+can0 ={       
+    "latest" : {
             "avg_rpm": 0,
             "avg_voltage": 0,
             "avg_power": 0,
@@ -54,16 +55,16 @@ def create_can0_data(elapsed_time: float) -> dict:
 
     return {
         "latest" : {
-            "avg_rpm": round((rpm_left + rpm_right) / 2, 2),
-            "avg_voltage": round(voltage, 2),
-            "avg_power": round((power_left + power_right) / 2, 2),
-            "speed": round(speed, 2),
+            "avg_rpm": int(round((rpm_left + rpm_right) / 2, 2)),
+            "avg_voltage": int(round(voltage, 2)),
+            "avg_power": int(round((power_left + power_right) / 2, 2)),
+            "speed": int(round(speed, 2)),
 
-            "power_left": round(power_left, 2),
-            "power_right": round(power_right, 2),
+            "power_left": int(round(power_left, 2)),
+            "power_right": int(round(power_right, 2)),
 
-            "current_left": round(current_left, 2),
-            "current_right": round(current_right, 2),
+            "current_left": int(round(current_left, 2)),
+            "current_right": int(round(current_right, 2)),
 
             "rpm_left": round(rpm_left, 2),
             "rpm_right": round(rpm_right, 2),
